refactor: 이메일 검증 및 에러 메세지 관리 기능 모듈화
-기존 로직에서 이메일 형식 검증 관련 함수를 utils 폴더의
formValidation.js로 분리
-에러 메세지 생성 및 삭제 관련 함수를 components 폴더의
errorMessage.js로 분리
프로젝트의 가독성과 관리 효율성 개선
