+++ signin.js
@@ -1,10 +1,13 @@
-import { emailCheck } from "./utils/formValidation";
+import { emailCheck } from "./utils/formValidation.js";
+import {
+  emailErrorMessage,
+  passwordErrorMessage,
+} from "./components/errorMessage.js";
 
-//로그인 버튼 조건부 활성화하기
 const loginBtn = document.querySelector(".submit");
 const form = document.querySelector(".form-container");
 
-//키보드 이벤트에 반응해서 로그인 활성화 상태 변경하기
+//로그인 활성화 상태 변경하기
 form.addEventListener("keyup", e => {
   if (
     emailInput.value === "" ||
@@ -19,44 +22,6 @@ form.addEventListener("keyup", e => {
     loginBtn.style.backgroundColor = "#3692ff";
   }
 });
-
-//에러 메세지 추가 함수
-const emailErrorMessage = (e, text) => {
-  //빈 값일 경우 빨간 테두리 추가
-  e.target.style.border = "1.5px solid red";
-  //하단에 메세지 추가
-  const message = document.createElement("span");
-  message.textContent = `${text}`;
-  message.style.color = "#F74747";
-  message.style.fontSize = "15px";
-  message.style.fontFamily = "Pretendard";
-  message.style.fontWeight = "600";
-  message.style.marginTop = "-16px";
-  message.style.marginBottom = "24px";
-  message.style.marginLeft = "16px";
-  message.classList.add("email-error-message");
-  message.classList.add("error");
-  e.target.after(message);
-};
-
-//패스워드 에러 메세지 추가 함수
-const passwordErrorMessage = (e, text, classname) => {
-  //빈 값일 경우 빨간 테두리 추가
-  e.target.style.border = "1.5px solid red";
-  //하단에 메세지 추가
-  const message = document.createElement("span");
-  message.textContent = `${text}`;
-  message.style.color = "#F74747";
-  message.style.fontSize = "15px";
-  message.style.fontFamily = "Pretendard";
-  message.style.fontWeight = "600";
-  message.style.marginTop = "-16px";
-  message.style.marginBottom = "24px";
-  message.style.marginLeft = "16px";
-  message.classList.add(classname);
-  message.classList.add("error");
-  e.target.after(message);
-};
 
 const emailInput = document.querySelector("#email");
 //이메일 포커스 아웃 이벤트
