+++ signup.js
@@ -1,4 +1,9 @@
-import { emailCheck } from "./utils/formValidation";
+import { emailCheck } from "./utils/formValidation.js";
+import {
+  emailErrorMessage,
+  passwordErrorMessage,
+  nicknameErrorMessage,
+} from "./components/errorMessage.js";
 
 const loginBtn = document.querySelector(".submit");
 const form = document.querySelector(".form-container");
@@ -21,44 +26,6 @@ form.addEventListener("keyup", e => {
     loginBtn.style.backgroundColor = "#9ca3af";
   }
 });
-
-//에러 메세지 추가 함수
-const emailErrorMessage = (e, text) => {
-  //빈 값일 경우 빨간 테두리 추가
-  e.target.style.border = "1.5px solid red";
-  //하단에 메세지 추가
-  const message = document.createElement("span");
-  message.textContent = `${text}`;
-  message.style.color = "#F74747";
-  message.style.fontSize = "15px";
-  message.style.fontFamily = "Pretendard";
-  message.style.fontWeight = "600";
-  message.style.marginTop = "-16px";
-  message.style.marginBottom = "24px";
-  message.style.marginLeft = "16px";
-  message.classList.add("email-error-message");
-  message.classList.add("error");
-  e.target.after(message);
-};
-
-//패스워드 에러 메세지 추가 함수
-const passwordErrorMessage = (e, text, classname) => {
-  //빈 값일 경우 빨간 테두리 추가
-  e.target.style.border = "1.5px solid red";
-  //하단에 메세지 추가
-  const message = document.createElement("span");
-  message.textContent = `${text}`;
-  message.style.color = "#F74747";
-  message.style.fontSize = "15px";
-  message.style.fontFamily = "Pretendard";
-  message.style.fontWeight = "600";
-  message.style.marginTop = "-16px";
-  message.style.marginBottom = "24px";
-  message.style.marginLeft = "16px";
-  message.classList.add(classname);
-  message.classList.add("error");
-  e.target.after(message);
-};
 
 const emailInput = document.querySelector("#email");
 //이메일 포커스 아웃 이벤트
@@ -108,25 +75,6 @@ passwordInput.addEventListener("focusin", function (e) {
     message.remove();
   }
 });
-
-//에러 메세지 추가 함수
-const nicknameErrorMessage = (e, text) => {
-  //빈 값일 경우 빨간 테두리 추가
-  e.target.style.border = "1.5px solid red";
-  //하단에 메세지 추가
-  const message = document.createElement("span");
-  message.textContent = `${text}`;
-  message.style.color = "#F74747";
-  message.style.fontSize = "15px";
-  message.style.fontFamily = "Pretendard";
-  message.style.fontWeight = "600";
-  message.style.marginTop = "-16px";
-  message.style.marginBottom = "24px";
-  message.style.marginLeft = "16px";
-  message.classList.add("nickname-error-message");
-  message.classList.add("error");
-  e.target.after(message);
-};
 
 //닉네임 포커스 아웃 이벤트
 const nicknameInput = document.querySelector("#nickname");
